Fix `.htaccess-lando` not being read when present.
Apache looks for `.htaccess` files in the order listed in `AccessFileName`,
reading only the first one it finds. Lando is supposed to change that
list, but it isn't working. See:
https://github.com/lando/lando/issues/2277
This project needs the updated `AccessFileName` list so the `.htaccess`
file on the server can include Shibboleth statements that will cause
errors in Lando, while the local, Lando-specific file, excludes them.